refactor: 공지사항 등록 및 삭제 로직 수정, 테스트 코드 추가
