+++ src/main/java/com/kgu/studywithme/study/domain/notice/Notice.java
@@ -37,7 +37,7 @@ public class Notice extends BaseEntity {
     @JoinColumn(name = "writer_id", referencedColumnName = "id", nullable = false)
     private Member writer;
 
-    @OneToMany(mappedBy = "notice", cascade = {CascadeType.PERSIST})
+    @OneToMany(mappedBy = "notice", cascade = CascadeType.PERSIST)
     private List<Comment> comments = new ArrayList<>();
 
     @Builder
