+++ src/main/java/com/kgu/studywithme/study/domain/notice/NoticeRepository.java
@@ -3,4 +3,5 @@ package com.kgu.studywithme.study.domain.notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
+    boolean existsByIdAndWriterId(Long noticeId, Long writerId);
 }
