+++ src/main/java/com/kgu/studywithme/study/domain/notice/comment/CommentRepository.java
@@ -3,4 +3,5 @@ package com.kgu.studywithme.study.domain.notice.comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
+    void deleteByNoticeId(Long noticeId);
 }
