+++ src/test/java/com/kgu/studywithme/common/ServiceTest.java
@@ -6,6 +6,7 @@ import com.kgu.studywithme.favorite.domain.FavoriteRepository;
 import com.kgu.studywithme.member.domain.MemberRepository;
 import com.kgu.studywithme.study.domain.StudyRepository;
 import com.kgu.studywithme.study.domain.notice.NoticeRepository;
+import com.kgu.studywithme.study.domain.notice.comment.CommentRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -34,6 +35,9 @@ public class ServiceTest {
 
     @Autowired
     protected NoticeRepository noticeRepository;
+
+    @Autowired
+    protected CommentRepository commentRepository;
 
     @AfterEach
     void clearDatabase() {
