+++ src/test/java/com/kgu/studywithme/study/service/StudyValidatorTest.java
@@ -3,8 +3,10 @@ package com.kgu.studywithme.study.service;
 import com.kgu.studywithme.common.ServiceTest;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
 import com.kgu.studywithme.member.domain.Member;
+import com.kgu.studywithme.member.exception.MemberErrorCode;
 import com.kgu.studywithme.study.domain.Study;
 import com.kgu.studywithme.study.domain.StudyName;
+import com.kgu.studywithme.study.domain.notice.Notice;
 import com.kgu.studywithme.study.exception.StudyErrorCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -40,5 +42,27 @@ class StudyValidatorTest extends ServiceTest {
                 .isInstanceOf(StudyWithMeException.class)
                 .hasMessage(StudyErrorCode.DUPLICATE_NAME.getMessage());
         assertDoesNotThrow(() -> studyValidator.validateName(diff));
+    }
+    
+    @Test
+    @DisplayName("스터디 팀장에 대한 검증을 진행한다")
+    void validateHost() {
+        assertThatThrownBy(() -> studyValidator.validateHost(study.getId(), host.getId() + 100L))
+                .isInstanceOf(StudyWithMeException.class)
+                .hasMessage(StudyErrorCode.MEMBER_IS_NOT_HOST.getMessage());
+        assertDoesNotThrow(() -> studyValidator.validateHost(study.getId(), host.getId()));
+    }
+    
+    @Test
+    @DisplayName("스터디 공지사항 작성자에 대한 검증을 진행한다")
+    void validateNoticeWriter() {
+        // given
+        Notice notice = noticeRepository.save(Notice.writeNotice(study, "공지사항", "내용"));
+
+        // when - then
+        assertThatThrownBy(() -> studyValidator.validateNoticeWriter(notice.getId(), host.getId() + 100L))
+                .isInstanceOf(StudyWithMeException.class)
+                .hasMessage(MemberErrorCode.MEMBER_IS_NOT_WRITER.getMessage());
+        assertDoesNotThrow(() -> studyValidator.validateNoticeWriter(notice.getId(), host.getId()));
     }
 }
